fix:修复重构后前后端端单元测试问题

+++ backend/src/main/java/com/spec/kit/exam/system/controller/BaseController.java
@@ -10,9 +10,9 @@ import java.util.function.Supplier;
 public abstract class BaseController {
 
     protected Map<String, Object> buildStatistics(
-        Supplier<Long> totalSupplier,
-        Supplier<Long> activeSupplier,
-        Supplier<Long> inactiveSupplier
+        Supplier<Integer> totalSupplier,
+        Supplier<Integer> activeSupplier,
+        Supplier<Integer> inactiveSupplier
     ) {
         Map<String, Object> statistics = new HashMap<>();
         statistics.put("total", totalSupplier.get());
